% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_congress}
\alias{add_congress}
\title{Add congress id to a table}
\usage{
add_congress(x, ...)
}
\arguments{
\item{x}{table}

\item{...}{other arguments for \code{\link[USAboundaries]{us_congress}}}
}
\value{
the inout table with congress added
}
\description{
Add congress id to a table
}
