% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_county}
\alias{add_county}
\title{Add county geoid to a table}
\usage{
add_county(x, ...)
}
\arguments{
\item{x}{sf table}

\item{...}{other arguments for \code{\link[USAboundaries]{us_counties}}}
}
\value{
the input table with county added
}
\description{
Add county geoid to a table
}
