% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch the data for one or more yea
rs as a tibble}
\usage{
fetch_data(year = years())
}
\arguments{
\item{year}{integer one or more years 2013-present}
}
\value{
tibble
}
\description{
Fetch the data for one or more yea
rs as a tibble
}
