% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_mst}
\alias{fetch_mst}
\title{Fetch the MST data for one or more years as a tibble}
\usage{
fetch_mst(
  year = current_year(),
  save_data = TRUE,
  save_path = get_path(),
  geocode = TRUE
)
}
\arguments{
\item{year}{integer one or more years 2013-present}

\item{save_data}{logical, if TRUE, then save the data}

\item{save_path}{character the path to save to}

\item{geocode}{logical if TRUE geocode the locations with lon, lat}
}
\value{
tibble
}
\description{
Fetch the MST data for one or more years as a tibble
}
