% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{geo_code}
\alias{geo_code}
\title{Geocode (determine lon,lat in this case) for one or more locations}
\usage{
geo_code(x)
}
\arguments{
\item{x}{tibble as returned by \code{\link{fetch_data}}}
}
\value{
an updated version of the input with lon and lat assigned best geocodes
}
\description{
Geocode (determine lon,lat in this case) for one or more locations
}
