% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_data_url}
\alias{get_data_url}
\title{Retrieve the data url for a given year}
\usage{
get_data_url(
  year = 2022,
  base_url = "https://mass-shooting-tracker-data.s3.us-east-2.amazonaws.com"
)
}
\arguments{
\item{year}{integer, one or more 4 digit years 2013-present}

\item{base_url}{character, https://massshootingtracker.site/ base data url}
}
\value{
the full url to a year's data
}
\description{
Retrieve the data url for a given year
}
