% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_path}
\alias{get_path}
\title{Construct a path to locally stored data}
\usage{
get_path(..., root = rappdirs::user_data_dir("columbine"))
}
\arguments{
\item{...}{path segments as a character vector}

\item{root}{character, the root path}
}
\value{
a constructed file path
}
\description{
Construct a path to locally stored data
}
