% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{leaflet_mst}
\alias{leaflet_mst}
\title{Draw a leaflet map}
\usage{
leaflet_mst(x = read_mst(), what = c("killed", "wounded", "shot")[3])
}
\arguments{
\item{x}{sf table}

\item{what}{character one or more of 'killed', 'wounded' or 'shot' (default)}
}
\value{
leaflet object
}
\description{
Draw a leaflet map
}
