% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_mst}
\alias{list_mst}
\title{Retrieve a listing of available files for MST}
\usage{
list_mst(years = seq_years(), path = get_path())
}
\arguments{
\item{years}{one or more years to read}

\item{path}{character the data path}
}
\value{
character vector of zero or more files
}
\description{
Retrieve a listing of available files for MST
}
