% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_years}
\alias{list_years}
\title{Retrieve a listing of available files}
\usage{
list_years(path = get_path())
}
\arguments{
\item{path}{character the data path}
}
\value{
character vector of zero or more files
}
\description{
Retrieve a listing of available files
}
