% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_data}
\alias{read_data}
\title{Read one or more years of data}
\usage{
read_data(
  years = seq_years(),
  path = get_path(),
  keep = c("date", "killed", "wounded", "city", "state", "lat", "lon"),
  form = c("data.frame", "sf")[1]
)
}
\arguments{
\item{years}{one or more years to read}

\item{path}{character, the data path}

\item{keep}{character, one or maore variables to keep, the default includes
"date", "killed", "wounded", "city", "state", "lat", "lon".  Specify "all"
to retrieve all fields.  Not compatible with \code{form = "sf"}.}

\item{form}{character, one of 'data.frame' or 'sf'}
}
\value{
data frame, possibly sf POINT table
}
\description{
Read one or more years of data
}
