% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_mst}
\alias{read_mst}
\title{Read one or more years of MST data}
\usage{
read_mst(
  years = seq_years(),
  path = get_path(),
  keep = c("date", "killed", "wounded", "city", "county", "congress", "state", "lat",
    "lon"),
  form = c("data.frame", "sf")[2],
  crop = sf::st_as_sfc(sf::st_bbox(c(xmin = -180, ymin = 0, xmax = 0, ymax = 80), crs =
    4326)),
  complete = tolower(form[1]) == "sf"
)
}
\arguments{
\item{years}{one or more years to read, by default all available}

\item{path}{character, the data path}

\item{keep}{character, one or maore variables to keep, the default includes
"date", "killed", "wounded", "city", "county_id", "congress_id", "state", "lon", "lat".
 Specify "all" to retrieve all fields.}

\item{form}{character, one of 'data.frame' or 'sf'}

\item{crop}{sfc or sf object defining a bounding box, ignored if \code{form} is not
'sf'.  Set to NULL to skip cropping.}

\item{complete}{logical, if TRUE remove records without lon,lat.  Ignored if
\code{form} is 'sf' in which case only complete cases of lon and lat are permitted.}
}
\value{
data frame, possibly sf POINT table
}
\description{
Read one or more years of MST data
}
