% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{seq_years}
\alias{seq_years}
\title{Create a sequence of years, 2013 to present}
\usage{
seq_years(start = 2013, end = current_year())
}
\arguments{
\item{start}{integer, the first year in the sequence}

\item{end}{the last year in the sequence}
}
\value{
integer sequence of years
}
\description{
Create a sequence of years, 2013 to present
}
